+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatPackageInformation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32a8fa04d86af7248aa35a2c3fd5c79b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
